Detailed API

EISAnalysis.CPE — Type

CPE(Q,n,ω,Z)

Struct for storing CPE information.

#Attributes

Q: CP capacitance

n: CP phase angle

ω: EIS frequencies

Z: EIS impedance

EISAnalysis.Capacitor — Type

Capacitor(C,ω,Z)

Struct for storing resistor information.

Attributes

C: Capacitance

ω: EIS frequencies

Z: EIS impedance

EISAnalysis.Circuit — Type

Circuit

Struct for storing information about circuits

Attributes

ω: EIS frequencies

Z: EIS Impedances

elements: List of circuit elements. Also stores paramter information.

operators: List of operators (-,/) between elements

order: Order of operations

subcircuits: List of subcircuits

EISAnalysis.Inductor — Type

Inductor(L,ω,Z)

Struct for storing inductor information.

Attributes

L: Inductance

ω: EIS frequencies

Z: EIS impedance

EISAnalysis.Resistor — Type

Resistor(R,ω,Z)

Struct for storing resistor information.

Attributes

R: Resistance

ω: EIS frequencies

Z: EIS impedance

EISAnalysis.Warburg — Type

Warburg(type,A,B,ω,Z)

Struct for storing Warburg information.

Can be either short or open, depending on type.

#Attributes

type: Warburg type ("short" or "open")

A: Warburg impedance parameter

B: Warburg exponential parameter

ω: EIS frequencies

Z: EIS impedance

Base.:* — Method

Base.*(a::Real,b::CircuitElement)

Mutates the impedance parameter of CircuitElements.

Arguments

a::Real: Impedance parameter value

b::CircuiElement: Circuit element

#Examples

using EISAnalysis
eval(initialize())
twor = 2r
twor.R == 2*r.R

# output

true

Base.:- — Method

-(a::Union{Circuit,CircuitElement},b::Union{Circuit,CircuitElement})

Holds the inputs in series and generates a Circuit.

Operates over Circuits and CircuitElements.

Arguments

a::Union{Circuit,CircuitElement}: circuit or circuit element to hold in series with b

b::Union{Circuit,CircuitElement}: circuit or circuit element to hold in series with a

#Examples

using EISAnalysis
eval(initialize())
circuit1 = r-c
circuit2 = circuit1-c
circuit2.Z == (r-c-c).Z == (r-0.5c).Z

# output

true

Base.:/ — Method

Base./(a::Union{Circuit,CircuitElement},b::Union{Circuit,CircuitElement})

Holds the inputs in parallel and generates a Circuit.

Operates over Circuits and CircuitElements.

Arguments

a::Union{Circuit,CircuitElement}: circuit or circuit element to hold in parallel with b

b::Union{Circuit,CircuitElement}: circuit or circuit element to hold in parallel with a

#Examples

using EISAnalysis
eval(initialize())
circuit1 = r/c
circuit2 = circuit1/c;
circuit2.Z == ((r/c)/c).Z

# output

true

Base.:^ — Method

Base.^(a::CircuitElement,b::Real)

Mutates the exponent parameter of CPE's and Warburgs.

Arguments

a::CircuiElement: circuit or circuit element to hold in series with a

b::Real: Exponential parameter value

#Examples

using EISAnalysis
eval(initialize())
circuit = q-wo 
print_circuit(circuit)
circuit2 = q^0.6-wo^5;
print_circuit(circuit2)

# output

1.0 * q ^ 0.8
1.0 * wo ^ 1.0
1.0 * q ^ 0.6
1.0 * wo ^ 5.0

Base.:~ — Method

Base.~(a::Union{CircuitElement,Circuit},ω::Vector{Float64})

Maps the frequencies over which impedance is calculated to the desired frequency.

Operates over Circuits and CircuitElements.

Arguments

-a: Circuit element or circuit -ω: Frequencies to map to

Examples

ω = [0.1,1,10]
using EISAnalysis
eval(initialize());
circuit = r/c ~ω; println(circuit.ω,circuit.Z)

# output

Real[0.1, 1.0, 10.0]ComplexF64[0.9900990099009901 - 0.09900990099009901im, 0.5 - 0.5im, 0.009900990099009903 - 0.09900990099009901im]

EISAnalysis.build_Z_matrices — Method

build_Z_matrices(ω_in,ω_out)

Builds the matrices (real + imaginary) through which impedance is calculated.

See compute_drt, optimize_lambda, and tune_τ

#Attributes

ω_in: The input EIS frequencies

ω_out: he desired output frequencies for computing DRT

EISAnalysis.circuit_fit — Method

circuit_fit(circuit, ω_data,Z_data)

Main function for fitting a Circuit to EIS data.

The initial guess is passed implicitly through the circuit definition. E.g. randles_circuit = 0.23r-(r-0.025wo^80)/0.2c # p0 = [0.23,1.0,(0.025,80),0.2]

Attributes

circuit: Circuit for fitting

ω_data: EIS frequencies

Z_data: EIS impedance

EISAnalysis.compute_drt — Method

compute_drt(ω_exp,Z_exp;ppd,showplot,rtol,regularization)

Main function for computing the DRT from input EIS data.

Attributes

ω_exp: Input EIS frequency

Z_exp: Input EIS Impedance

ppd=7: Points-per-decade in output τ range for computing DRT

showplot=false: Whether or not to plot DRT results

rtol=1e-03: The desired relative tolerance. Function outputs a warning if the fit result is above this

regularization=false: Whether or no to employ the regularization method

Examples

using EISAnalysis, Statistics
eval(initialize())
ω_exp, Z_exp = (r/q).ω, (r/q).Z #replace this with real data
fit = compute_drt(ω_exp,Z_exp;showplot = false,regularization = true)

EISAnalysis.drt_Z — Method

drt_Z(X_r,X_i,p)

Uses Z matrices to compute impedance for least-squares fit.

Attributes

X_r: The real Z matrix for computing impedance

X_i: The imaginary Z matrix for computing impedance

p: The parameter list

EISAnalysis.drt_Z_regular — Method

drt_Z_regular(X_r,X_i,Y_r,Y_i,λ,p_reg)

See drt_Z. Incorporates regularization in least-squares minimization.

Since LsqFit minimizes Σ|Xᵢ-Yᵢ|², this function generates A such that |Aᵢ-Yᵢ|² = |Xᵢ-Yᵢ|² + (λ/N)|p|² Aᵢ = Yᵢ+√(|Xᵢ-Yᵢ|²+(λ/N)|p|²) where N is the length of the vectors. Since the Real component carries the extra parameter (R0), two different functionsare defined for the real and imaginary components.Used for regularization method of computing DRT.

Attributes

X_r: The real Z matrix for computing impedance

X_i: The imaginary Z matrix for computing impedance

Y_r: The real component of input impedance data

Y_i: The imaginary component of input impedance data

λ: The regularization hyperparameter

p_reg: The parameter list

EISAnalysis.drt_Zimag_regular — Method

drt_Zimag_regular(X_i,Y_i,λ,p_reg)

See drt_Z_regular. Outputs the imaginary component.

The only difference with drtZrealregular() is ignoring the series resistance. Used for crossvalidation in optimize_lambda.

Attributes

X_i: The imaginary Z matrix for computing impedance

Y_i: The imaginary component of input impedance data

λ: The regularization hyperparameter

p_reg: The parameter list

EISAnalysis.drt_Zreal_regular — Method

drt_Zreal_regular(X_r,Y_r,λ,p_reg)

See drt_Z_regular. Outputs the real component.

Used for crossvalidation in optimize_lambda.

Attributes

X_r: The real Z matrix for computing impedance

Y_r: The real component of input impedance data

λ: The regularization hyperparameter

p_reg: The parameter list

EISAnalysis.evaluate_Z — Method

evaluate_Z(f_r,f_c)

Computes elements for Z_matrices. The function based on impedance of an RC Circuit

Arguments

f_r: The Z_matrix row frequency

f_c: The Z_matrix column frequency

EISAnalysis.get_params — Method

get_params(a::Union{CircuitElement,Circuit})

Gets the parameters for elements in a circuit.

Attributes

a::Union{CircuitElement,Circuit}: The circuit or circuit element

Examples

using EISAnalysis
eval(initialize())
randles_circuit = 0.23r-(r-0.025wo^80)/0.2q
get_params(randles_circuit)

# output

4-element Vector{Any}:
 0.23
 1.0
  (0.025, 80.0)
  (0.2, 0.8)

EISAnalysis.get_subcircuit — Method

get_subcircuit(subelements,suboperators,suborder)

Creates a circuit from a subcircuit.

Used in rebuild. Currently a bit sloppy.

Attributes

subelements: elements of subcircuit

suboperators: operators of subcircuit

suborder: operation order of subcircuit

EISAnalysis.get_symbol — Method

get_symbol(a::CircuitElement)

Creates the symbol (technically expressions) for a circuit element, using globally defined variables.

Used for generating the elements field of a circuit.

Attributes

a::CircuitElement: The circuit element

EISAnalysis.initialize — Method

initialize()

Generates all the circuit elements for ease of use in building circuits.

It adds the following variables to your environment r = Resistor() c = Capacitor() l = Inductor() q = CPE() wo = Warburg("open") ws = Warburg("short") From here you can quickly build circuits and adjust the parameters directly using the overloaded * and ^ operators as desired

Examples

using EISAnalysis
eval(initialize())
print(r.R)

# output

1.0

EISAnalysis.optimize_lambda — Method

optimize_lambda(ω_exp,Z_exp,τ)

Generates the optimal regularization hyperparameter for compute_drt.

Attributes

ω_exp: Input EIS frequency

Z_exp: Input EIS Impedance

τ: Desired relaxation times for computing DRT: generated in compute_drt

EISAnalysis.plot_Nyquist! — Method

plot_Nyquist!(plt,a;label)

Adds circuits to pre-existing Nyquist plot

Arguments

plt: The input Nyquist plot to manipulate

a::Circuit: The circuits to add to the Nyquist plot

label: kwarg label for plot

Examples

julia> using EISAnalysis
julia> eval(initialize());
julia> circuit1 = r-r/c;
julia> circuit2 = r-r/q;
julia> plot_Nyquist(circuit1,circuit2;label= ["R-C Circuit","R-CPE Circuit"]);

EISAnalysis.plot_Nyquist — Method

plot_Nyquist(a;label)

Creates a Nyquist plot

Arguments

a::Circuit: The circuits to add to the Nyquist plot

label: kwarg label for plot

Examples

julia> using EISAnalysis
julia> eval(initialize());
julia> circuit1 = r-r/c;
julia> circuit2 = r-r/q;
julia> plot_Nyquist(circuit1,circuit2;label= ["R-C Circuit","R-CPE Circuit"]);

EISAnalysis.print_circuit — Method

print_circuit(circuit)

Prints the elements of a circuit along with its parameters

Arguments

circuit::Circuit: The circuit being printed

Examples

using EISAnalysis
eval(initialize())
randles_circuit = 0.23r-r/0.2q
print_circuit(randles_circuit)

# output 

0.23r
1.0r
0.2 * q ^ 0.8

EISAnalysis.rebuild — Method

rebuild(circuit::Circuit)

Main function for recalculating a circuit's impedance.

Used after mutating a circuit through either ~ or set_params

Attributes

circuit: Mutated circuit to be rebuilt

EISAnalysis.regularizer — Method

regularizer(p,λ)

Computes the regularization of parameters. Used in regularization method of computing DRT.

See drt_Z_regular,drt_Zimag_regular, and drt_Zreal_regular

#Arguments

p: The parameter list

λ: Regularization hyperparameter

EISAnalysis.set_params — Method

set_params(a::Union{CircuitElement,Circuit})

Sets the parameters for elements in a circuit.

Currently a bit sloppy.Used in circuit_fit

Attributes

a::Union{CircuitElement,Circuit}: The circuit or circuit element

p: The parameter list. Needs to carry tuples for elements with two parameters

Examples

using EISAnalysis
eval(initialize());
circuit = r-r/q;
p = [0.5,2,(0.5,0.9)]
updated_circuit = EISAnalysis.set_params(circuit,p)
print_circuit(updated_circuit)

# output

0.5r
2.0r
0.5 * q ^ 0.9

EISAnalysis.tune_τ — Method

tune_τ(ω_exp,Z_exp;ppd,tol)

Generates the optimal τ range for computing DRT.

Currently incomplete. Would be used in compute_drt

Attributes

ω_exp: Input EIS frequency

Z_exp: Input EIS Impedance

ppd: Points-per-decade in output τ range for computing DRT

tol=1e-03: Tolerance for finding τ bounds where DRT impedance is sufficiently small

EISAnalysis.unflatten_parameters — Method

unflatten_parameters(pflat,tuples)

Shapes parameter list to feed into Circuit. See circuit_fit

LsqFit.curvefit requires a parameter list that is flat, but setparams may require a parameter list with tuples. This function unflattens parameter list to feed to set_params.

Attributes

pflat: Flat parameter list

tuples: List of tuple indices for unflattened parameter list